requirements.txt
asgiref==3.2.10
Django==3.1.2
pytz==2020.1
sqlparse==0.4.1
psycopg2-binary==2.8.6
django-allauth==0.43.0
django-extensions==3.0.9
Werkzeug==1.0.1
pyOpenSSL==19.1.0
django-crispy-forms==1.9.2
selenium
webdriver-manager